refactor move components to view
